fix compiler warnings for tests

+++ Makefile
@@ -53,7 +53,7 @@ STATIC_MODULES+=fanotify fork inotify netlink_raw nl_route_raw uevents socket_ra
 DYN_MODULES+=avahi can curl evdev evloop_qt libnl libvirt netlink_ge nf_queue \
 	pulseaudio readline sdbus sip udev v4l xcb_randr sdbus_notifications
 
-LAYER2_MODULES?=dynamic_evdev forked_curl netif
+LAYER2_MODULES?=dynamic_evdev forked_curl netif netconf
 
 DEFAULT_EVLOOP=epoll
 
@@ -145,13 +145,13 @@ dllist.o: llist.c
 $(SHAREDLIB): LDLIBS+=$(foreach o,$(STATIC_TOOLS) $(STATIC_MODULES), $(LDLIBS_$(o)) )
 $(SHAREDLIB): $(OBJS)
 	$(CC) -shared -o $@.$(SO_VERSION) $(OBJS) $(LDFLAGS) $(LDLIBS) -Wl,-soname,$@.$(SO_VERSION)
-	ln -s $@.$(SO_VERSION) $@
+	[ -e $@ ] || ln -s $@.$(SO_VERSION) $@
 
 # libcrtx_%.so: $(OBJS_$(patsubst libcrtx_%.so,%,$@))
 libcrtx_%.so: LDLIBS+=$(LDLIBS_$(patsubst libcrtx_%.so,%,$@)) $(foreach d,$(DEPS_$(patsubst libcrtx_%.so,%,$@)),$(if $(findstring $(d),$(DYN_MODULES)),$(LDLIBS_$(d))) )
 libcrtx_%.so: %.o $(SHAREDLIB)
 	$(CC) -shared $(LDFLAGS) $< -o $@.$(SO_VERSION) -L. $(LDLIBS) -lcrtx -Wl,-soname,$@.$(SO_VERSION)
-	ln -s $@.$(SO_VERSION) $@
+	[ -e $@ ] || ln -s $@.$(SO_VERSION) $@
 
 plugins: $(SHAREDLIB) $(DYN_MODULES_LIST)
